requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
flask-cors
mecab-python3
gensim
